+++ x8086NetEmu/Chipset/PIC8259.vb
@@ -115,7 +115,7 @@
         If master IsNot Nothing Then UpdateSlaveOutput()
     End Sub
 
-    Public Overrides Function [In](port As UInt32) As UInt16
+    Public Overrides Function [In](port As UInt16) As Byte
         If (port And 1) = 0 Then
             ' A0 = 0
             If pollMode Then
@@ -129,7 +129,7 @@
         End If
     End Function
 
-    Public Overrides Sub Out(port As UInt32, value As UInt16)
+    Public Overrides Sub Out(port As UInt16, value As Byte)
         If (port And 1) = 0 Then
             ' A0 = 0
             If (value And &H10) <> 0 Then
